Move JUnit4-specific classes down into junit4 package
