+++ subprojects/testing-jvm/src/integTest/groovy/org/gradle/testing/junit/junit4/JUnit4ClassLevelFilteringIntegrationTest.groovy
@@ -17,7 +17,6 @@
 package org.gradle.testing.junit.junit4
 
 import org.gradle.integtests.fixtures.TargetCoverage
-import org.gradle.testing.junit.AbstractJUnit4ClassLevelFilteringIntegrationTest
 
 import static org.gradle.testing.fixture.JUnitCoverage.LARGE_COVERAGE
